requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
docutils==0.16
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
jeepney==0.4.3
keyring==21.2.1
multidict==4.7.6
packaging==20.4
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dotenv==0.13.0
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
six==1.15.0
tqdm==4.46.0
twine==3.1.1
twitchio==1.1.0
typing-extensions==3.7.4.2
urllib3==1.25.9
webencodings==0.5.1
websockets==8.1
yarl==1.4.2
zipp==3.1.0
